@@ -1,47 +1,42 @@
 {
   "name": "vue-shop",
-  "version": "0.1.0",
+  "version": "0.0.0",
   "private": true,
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint"
+    "dev": "vite",
+    "build": "run-p type-check build-only",
+    "preview": "vite preview",
+    "build-only": "vite build",
+    "type-check": "vue-tsc --noEmit -p tsconfig.app.json --composite false",
+    "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs,.ts,.tsx,.cts,.mts --fix --ignore-path .gitignore",
+    "format": "prettier --write src/"
   },
   "dependencies": {
-    "axios": "^0.25.0",
-    "core-js": "^3.20.3",
-    "js-cookie": "^3.0.1",
+    "axios": "^1.4.0",
     "mitt": "^3.0.0",
-    "moment": "^2.29.1",
-    "register-service-worker": "^1.7.2",
-    "swiper": "^7.4.1",
-    "ts-md5": "^1.2.10",
-    "vue": "^3.2.28",
-    "vue-class-component": "^8.0.0-rc.1",
-    "vue-property-decorator": "^10.0.0-rc.3",
-    "vue-router": "^4.0.12",
-    "vuex": "^4.0.0",
-    "vuex-class": "^0.3.2",
-    "vuex-class-modules": "^1.3.0"
+    "moment": "^2.29.4",
+    "pinia": "^2.1.3",
+    "swiper": "^9.4.1",
+    "ts-md5": "^1.3.1",
+    "vue": "^3.3.4",
+    "vue-router": "^4.2.2"
   },
   "devDependencies": {
-    "@types/js-cookie": "^3.0.1",
-    "@typescript-eslint/eslint-plugin": "^5.10.0",
-    "@typescript-eslint/parser": "^5.10.0",
-    "@vue/cli-plugin-babel": "~5.0.0-beta.3",
-    "@vue/cli-plugin-eslint": "~5.0.0-beta.3",
-    "@vue/cli-plugin-pwa": "~5.0.0-beta.3",
-    "@vue/cli-plugin-router": "~5.0.0-beta.3",
-    "@vue/cli-plugin-typescript": "~5.0.0-beta.3",
-    "@vue/cli-plugin-vuex": "~5.0.0-beta.3",
-    "@vue/cli-service": "~5.0.0-beta.3",
-    "@vue/compiler-sfc": "^3.2.28",
-    "@vue/eslint-config-typescript": "^10.0.0",
-    "eslint": "^8.7.0",
-    "eslint-formatter-codeframe": "^7.32.1",
-    "eslint-plugin-vue": "^8.3.0",
-    "sass": "^1.49.0",
-    "sass-loader": "^12.4.0",
-    "typescript": "^4.5.5"
+    "@rushstack/eslint-patch": "^1.2.0",
+    "@tsconfig/node18": "^2.0.1",
+    "@types/node": "^20.3.1",
+    "@vitejs/plugin-vue": "^4.2.3",
+    "@vue/eslint-config-prettier": "^7.1.0",
+    "@vue/eslint-config-typescript": "^11.0.3",
+    "@vue/tsconfig": "^0.4.0",
+    "eslint": "^8.39.0",
+    "eslint-plugin-vue": "^9.11.0",
+    "npm-run-all": "^4.1.5",
+    "prettier": "^2.8.8",
+    "sass": "^1.63.6",
+    "typescript": "~5.1.3",
+    "vite": "^4.3.9",
+    "vite-plugin-pwa": "^0.16.4",
+    "vue-tsc": "^1.6.5"
   }
 }
