@@ -1,4 +1,4 @@
-import { IPage, IComment, ICommentSubtotal } from './model';
+import type { IPage, IComment, ICommentSubtotal } from './model';
 import { fetch } from '@/utils/http';
 
 export const getCommentList = (param: any) => fetch<IPage<IComment>>('shop/comment', param);
