@@ -33,78 +33,78 @@
 <script lang="ts">
 import { removeLocalStorage, getLocalStorage, setLocalStorage } from '@/utils';
 const KEYWORDS_HISTORY = 'KEYWORDS_HISTORY';
-import { Vue, Prop, Emit, Options } from 'vue-property-decorator';
 import SearchHeader from './SearchHeader.vue';
 import { getHotKeywords, getTips } from '@/api/product';
+import { defineComponent } from 'vue'
 
-@Options({
+export default defineComponent({
     components: {
-        SearchHeader,
+        SearchHeader
     },
-})
-export default class SearchBar extends Vue {
-    @Prop(String) public readonly value!: string;
-
-    public hotKeywords?: string[] = [];
-    public tipList?: string[] = [];
-    public historyList: string[] = [];
-    public isMini: boolean = false;
-
-    public created() {
+    props: {
+        value: String,
+    },
+    data() {
+        const hotKeywords: string[] = [];
+        const tipList: string[] = [];
+        const historyList: string[] = [];
+        return {
+            hotKeywords,
+            tipList,
+            historyList,
+            isMini: false
+        };
+    },
+    created() {
         this.isMini = Object.keys(this.$route.query).length > 0;
-        this.historyList = getLocalStorage<string[]>(KEYWORDS_HISTORY, true) || []
+        this.historyList = getLocalStorage<string[]>(KEYWORDS_HISTORY, true) as any || [];
         getHotKeywords().then(res => {
-            this.hotKeywords = res.data
-        })
-    }
-
-
-    public updateVal(val: string) {
-        this.$emit('input', val);
-        if (!val || val.length < 1) {
-            this.tipList = [];
+            this.hotKeywords = res.data as any;
+        });
+    },
+    methods: {
+        updateVal(val: string) {
+            this.$emit('input', val);
+            if (!val || val.length < 1) {
+                this.tipList = [];
+                this.isMini = false;
+            }
+        },
+        tapClearHistory() {
+            this.historyList = []
+            removeLocalStorage(KEYWORDS_HISTORY)
+        },
+        addHistory(keywords: string) {
+            if (this.historyList.indexOf(keywords) >= 0) {
+                return;
+            }
+            this.historyList.push(keywords);
+            if (this.historyList.length > 8) {
+                this.historyList.splice(8);
+            }
+            setLocalStorage(KEYWORDS_HISTORY, this.historyList)
+        },
+        onKeyUp() {
+            if (!this.value || this.value.length < 1) {
+                this.tipList = [];
+                return;
+            }
+            getTips(this.value).then(res => {
+                this.tipList = res.data as any;
+            });
+        },
+        tapSearch(keywords: string) {
+            if (!keywords || keywords.trim().length === 0) {
+                return;
+            }
+            this.addHistory(keywords);
+            this.$emit('search', keywords);
+            this.isMini = true;
+        },
+        tapFocus() {
+            this.$emit('focus');
             this.isMini = false;
         }
     }
-
-    public tapClearHistory() {
-        this.historyList = []
-        removeLocalStorage(KEYWORDS_HISTORY)
-    }
-
-    public addHistory(keywords: string) {
-        if (this.historyList.indexOf(keywords) >= 0) {
-            return;
-        }
-        this.historyList.push(keywords);
-        if (this.historyList.length > 8) {
-            this.historyList.splice(8);
-        }
-        setLocalStorage(KEYWORDS_HISTORY, this.historyList)
-    }
-
-    public onKeyUp(event: any) {
-        if (!this.value || this.value.length < 1) {
-            this.tipList = [];
-            return;
-        }
-        getTips(this.value).then(res => {
-            this.tipList = res.data
-        });
-    }
-
-    public tapSearch(keywords: string) {
-        if (!keywords || keywords.trim().length === 0) {
-            return;
-        }
-        this.addHistory(keywords);
-        this.$emit('search', keywords);
-        this.isMini = true;
-    }
-
-    @Emit('focus')
-    public tapFocus() {
-        this.isMini = false;
-    }
-}
+});
 </script>
