@@ -1,5 +1,5 @@
 <template>
-    <div class="swipe-row" :style="{left: left + 'px'}">
+    <div class="swipe-row" :style="{left: input.left + 'px'}">
         <div class="actions-left" ref="left">
             <slot name="left"></slot>
         </div>
@@ -16,126 +16,127 @@
         </div>
     </div>
 </template>
-<script lang="ts">
-import { Vue, Prop, Ref } from 'vue-property-decorator';
+<script setup lang="ts">
+import { getCurrentInstance, reactive, ref } from 'vue';
 
-export default class SwipeRow extends Vue {
-    @Prop([String, Array]) public readonly name!: string| string[];
-    @Prop([Number, String]) public readonly index!: number|string;
-    public oldLeft = 0;
-    public left = 0;
-    public startX = 0;
-    public isTouch = false;
-    @Ref('left') public readonly leftBox!: HTMLDivElement;
-    @Ref('right') public readonly rightBox!: HTMLDivElement;
+const emit = defineEmits(['click', 'remove']);
+const props = defineProps<{
+    name: string|string[],
+    index: number|string;
+}>();
 
-    public getLeftWidth(): number {
-        if (!this.leftBox) {
-            return 0;
-        }
-        return this.leftBox.clientWidth || this.leftBox.offsetWidth;
+const input = reactive({
+    oldLeft: 0,
+    left: 0,
+    startX: 0,
+    isTouch: false
+})
+
+const leftBox = ref<HTMLDivElement>();
+const rightBox = ref<HTMLDivElement>();
+
+function getLeftWidth(): number {
+    if (!leftBox.value) {
+        return 0;
     }
-
-    public getRightWidth(): number {
-        if (!this.rightBox) {
-            return 0;
-        }
-        return this.rightBox.clientWidth || this.rightBox.offsetWidth;
+    return leftBox.value.clientWidth || leftBox.value.offsetWidth;
+}
+function getRightWidth(): number {
+    if (!rightBox.value) {
+        return 0;
     }
+    return rightBox.value.clientWidth || rightBox.value.offsetWidth;
+}
+function tapRemove(item: any) {
+    emit('remove', item);
+}
 
-    public tapRemove(item: any) {
-        this.$emit('remove', item);
-    }
-
-    public touchStart(e: TouchEvent) {
-        this.resetOther();
-        this.oldLeft = this.left;
-        this.isTouch = false;
-        this.startX = e.targetTouches[0].clientX;
-    }
-
-    public touchMove(e: TouchEvent) {
-        this.isTouch = true;
-        const diff = e.targetTouches[0].clientX - this.startX;
-        if (this.oldLeft === 0) {
-            if (diff < 0) {
-                this.left = Math.max(diff, -this.getRightWidth());
-                return;
-            }
-            this.left = Math.min(diff, this.getLeftWidth());
-            return;
-        }
-        if (this.oldLeft > 0) {
-            if (diff > 0) {
-                return;
-            }
-            this.left =  Math.max(this.oldLeft + diff, 0);
-            return;
-        }
+function touchStart(e: TouchEvent) {
+    resetOther();
+    input.oldLeft = input.left;
+    input.isTouch = false;
+    input.startX = e.targetTouches[0].clientX;
+}
+function touchMove(e: TouchEvent) {
+    input.isTouch = true;
+    const diff = e.targetTouches[0].clientX - input.startX;
+    if (input.oldLeft === 0) {
         if (diff < 0) {
+            input.left = Math.max(diff, -getRightWidth());
             return;
         }
-        this.left = Math.min(this.oldLeft + diff, 0);
+        input.left = Math.min(diff, getLeftWidth());
+        return;
     }
-
-    public touchEnd() {
-        if (!this.isTouch) {
-            this.animation(this.left, 0);
-            this.$emit('click');
+    if (input.oldLeft > 0) {
+        if (diff > 0) {
             return;
         }
-        // const diff = e.changedTouches[0].clientX - this.startX;
-        if (this.left === 0) {
-            return;
-        }
-        if (this.left > 0) {
-            const width = this.getLeftWidth();
-            this.animation(this.left, this.left * 3 > width ? width : 0);
-            return;
-        }
-        const w = - this.getRightWidth();
-        this.animation(this.left, this.left * 3 < w ? w : 0);
+        input.left = Math.max(input.oldLeft + diff, 0);
+        return;
     }
+    if (diff < 0) {
+        return;
+    }
+    input.left = Math.min(input.oldLeft + diff, 0);
+}
+function touchEnd() {
+    if (!input.isTouch) {
+        animation(input.left, 0);
+        emit('click');
+        return;
+    }
+    // const diff = e.changedTouches[0].clientX - startX;
+    if (input.left === 0) {
+        return;
+    }
+    if (input.left > 0) {
+        const width = getLeftWidth();
+        animation(input.left, input.left * 3 > width ? width : 0);
+        return;
+    }
+    const w = - getRightWidth();
+    animation(input.left, input.left * 3 < w ? w : 0);
+}
 
-    public animation(
-        start: number, end: number, endHandle?: () => void) {
-        const diff = start > end ? -1 : 1;
-        let step = 1;
-        const handle = setInterval(() => {
-            start += (step ++) * diff;
-            if ((diff > 0 && start >= end) || (diff < 0 && start <= end)) {
-                clearInterval(handle);
-                this.left = end;
-                if (typeof endHandle === 'function') {
-                    endHandle();
-                }
-                return;
+function animation(
+    start: number, end: number, endHandle?: () => void) {
+    const diff = start > end ? -1 : 1;
+    let step = 1;
+    const handle = setInterval(() => {
+        start += (step ++) * diff;
+        if ((diff > 0 && start >= end) || (diff < 0 && start <= end)) {
+            clearInterval(handle);
+            input.left = end;
+            if (typeof endHandle === 'function') {
+                endHandle();
             }
-            this.left = start;
-        }, 16);
-    }
+            return;
+        }
+        input.left = start;
+    }, 16);
+}
 
-    public reset() {
-        if (this.left === 0) {
-            return;
-        }
-        this.animation(this.left, 0);
+function reset() {
+    if (input.left === 0) {
+        return;
     }
-
-    public resetOther() {
-        if (typeof this.index === 'undefined' || !this.$parent) {
-            return;
+    animation(input.left, 0);
+}
+function resetOther() {
+    const parent = getCurrentInstance()?.parent;
+    if (typeof props.index === 'undefined' || !parent) {
+        return;
+    }
+    const items: any[] = parent.refs.swiperow as any[];
+    if (!items || items.length < 1) {
+        return;
+    }
+    for (const item of items) {
+        if (item.index === props.index) {
+            continue;
         }
-        const items: SwipeRow[] = this.$parent.$refs.swiperow as SwipeRow[];
-        if (!items || items.length < 1) {
-            return;
-        }
-        for (const item of items) {
-            if (item.index === this.index) {
-                continue;
-            }
-            item.reset();
-        }
+        item.reset();
     }
 }
 </script>
