@@ -6,8 +6,10 @@
         </a>
     </footer>
 </template>
-<script lang="ts">
-import { Vue } from 'vue-property-decorator';
+<script setup lang="ts">
+import { useRouter } from 'vue-router';
+
+const router = useRouter();
 
 interface IMenu {
     name: string,
@@ -15,32 +17,29 @@ interface IMenu {
     url: string
 }
 
-export default class TabBar extends Vue {
-    public menus: IMenu[]  = [
-        {
-            name: '首页',
-            icon: 'iconfont fa-home',
-            url: 'home',
-        },
-        {
-            name: '分类',
-            icon: 'iconfont fa-table',
-            url: 'category',
-        },
-        {
-            name: '购物车',
-            icon: 'iconfont fa-cart',
-            url: 'cart',
-        },
-        {
-            name: '我的',
-            icon: 'iconfont fa-user',
-            url: 'member',
-        },
-    ];
-
-    public tapMenu(item: IMenu) {
-        this.$router.push({ name: item.url});
-    }
+const menus: IMenu[]  = [
+    {
+        name: '首页',
+        icon: 'iconfont fa-home',
+        url: 'home',
+    },
+    {
+        name: '分类',
+        icon: 'iconfont fa-table',
+        url: 'category',
+    },
+    {
+        name: '购物车',
+        icon: 'iconfont fa-cart',
+        url: 'cart',
+    },
+    {
+        name: '我的',
+        icon: 'iconfont fa-user',
+        url: 'member',
+    },
+];
+function tapMenu(item: IMenu) {
+    router.push({ name: item.url});
 }
 </script>
