@@ -1,18 +1,22 @@
-import { createApp } from 'vue';
-import App from './App.vue';
-import './registerServiceWorker';
-import router from './router';
-import store from './store';
+import './assets/css/theme.scss'
+
+import { createApp } from 'vue'
+import { createPinia } from 'pinia'
+
+const pinia = createPinia();
+
+import App from './App.vue'
+import router from './router'
 import emitter from './event';
 
-import 'swiper/scss';
-import './assets/iconfont/iconfont.css';
 
-import http from './utils/http';
-import title from './utils/title';
-
-createApp(App, {
+const app = createApp(App, {
     onscroll(e: Event) {
         emitter.emit('scroll', e);
     }
-}).use(http).use(title).use(store).use(router).mount('#app');
+})
+
+app.use(pinia)
+app.use(router)
+
+app.mount('#app')
